+++ core/src/ap/jpki.rs
@@ -2,8 +2,8 @@
 
 use std::rc::Rc;
 
-use crate::jpki::card::Card;
 use crate::nfc;
+use crate::Card;
 
 const DF_NAME: [u8; 10] = [0xD3, 0x92, 0xF0, 0x00, 0x26, 0x01, 0x00, 0x00, 0x00, 0x01];
 const EF_AUTH: [u8; 2] = [0x00, 0x17];
