--- core/src/jpki.rs
@@ -1,5 +1,0 @@
-mod card;
-
-pub mod ap;
-
-pub use card::Card;
